Added export of analysis

+++ win_build.py
@@ -4,9 +4,10 @@
 Created on Mon Apr 13 21:19:42 2026
 
 @author: carl
+
+Builds Windows installer using pyinstaller and Inno Setup.
 """
 from pathlib import Path
-import shutil
 import subprocess
 import sys
 import os
@@ -27,12 +28,15 @@ def run(cmd):
 def main():
     run([sys.executable, "-m", "PyInstaller", "--clean", "--noconfirm", "recipe/exhale.spec"])
 
-    zip_path = ZIP_BASE.with_suffix(ZIP_BASE.suffix + ".zip")
-    if zip_path.exists():
-        zip_path.unlink()
-    shutil.make_archive(str(ZIP_BASE), "zip", root_dir=DIST_DIR.parent,
-         base_dir=DIST_DIR.name)
-    print("Created", zip_path)
+    do_zip = False
+    if do_zip:
+        import shutil
+        zip_path = ZIP_BASE.with_suffix(ZIP_BASE.suffix + ".zip")
+        if zip_path.exists():
+            zip_path.unlink()
+        shutil.make_archive(str(ZIP_BASE), "zip", root_dir=DIST_DIR.parent,
+             base_dir=DIST_DIR.name)
+        print("Created", zip_path)
 
     # Inno Setup
     run([
